Rename enhanced consent module
